requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.8
cryptography==44.0.0
Deprecated==1.2.15
ecdsa==0.19.0
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jwt==1.3.1
limits==3.14.1
line_profiler==4.1.3
memory-profiler==0.61.0
packaging==24.2
pip==24.3.1
pluggy==1.5.0
psutil==6.1.0
pyasn1==0.6.1
pybreaker==1.2.0
pycparser==2.22
pydantic==2.10.2
pydantic_core
uvicorn
sqlalchemy
psycopg2-binary
requests
python-jose[cryptography]
python-decouple